Renamed package to include privacyfriendly. Added some describe text to the game activity.
