+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.secuso.example">
+    package="org.secuso.privacyfriendlyexample">
 
     <application
         android:allowBackup="true"
@@ -9,7 +9,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name="org.secuso.privacyfriendlyexample.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
@@ -20,44 +20,44 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".AboutActivity"
+            android:name="org.secuso.privacyfriendlyexample.AboutActivity"
             android:label="@string/about"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName="org.secuso.privacyfriendlyexample.MainActivity"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="org.secuso.example.MainActivity" />
+                android:value="org.secuso.privacyfriendlyexample.MainActivity" />
         </activity>
         <activity
-            android:name=".HelpActivity"
+            android:name="org.secuso.privacyfriendlyexample.HelpActivity"
             android:label="@string/help"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName="org.secuso.privacyfriendlyexample.MainActivity"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="org.secuso.example.MainActivity" />
+                android:value="org.secuso.privacyfriendlyexample.MainActivity" />
         </activity>
         <activity
-            android:name=".SettingsActivity"
+            android:name="org.secuso.privacyfriendlyexample.SettingsActivity"
             android:label="@string/title_activity_settings"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName="org.secuso.privacyfriendlyexample.MainActivity"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="org.secuso.example.MainActivity" />
+                android:value="org.secuso.privacyfriendlyexample.MainActivity" />
         </activity>
         <activity
-            android:name=".GameActivity"
+            android:name="org.secuso.privacyfriendlyexample.GameActivity"
             android:label="@string/title_activity_game"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName="org.secuso.privacyfriendlyexample.MainActivity"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="org.secuso.example.MainActivity" />
+                android:value="org.secuso.privacyfriendlyexample.MainActivity" />
         </activity>
     </application>
