+++ app/src/main/java/org/secuso/privacyfriendlyexample/AboutActivity.java
@@ -1,12 +1,6 @@
-package org.secuso.example;
+package org.secuso.privacyfriendlyexample;
 
-import android.graphics.Color;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.graphics.drawable.ColorDrawable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.MenuItem;
 
 /**
  * Created by yonjuni on 15.06.16.
