+++ app/src/main/res/layout/activity_about.xml
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
-        tools:context="org.secuso.example.MainActivity">
+        tools:context="org.secuso.privacyfriendlyexample.MainActivity">
 
         <android.support.design.widget.AppBarLayout
             android:layout_width="match_parent"
