+++ app/src/main/res/layout/activity_help.xml
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
-        tools:context="org.secuso.example.HelpActivity">
+        tools:context="org.secuso.privacyfriendlyexample.HelpActivity">
 
         <include layout="@layout/toolbar"/>
 
@@ -27,7 +27,7 @@
             android:orientation="vertical"
             tools:context=".HelpActivity"
             android:weightSum="1">
-            <fragment android:name="org.secuso.example.HelpActivity$HelpFragment"
+            <fragment android:name="org.secuso.privacyfriendlyexample.HelpActivity$HelpFragment"
                 android:id="@+id/help_fragment"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"/>
