+++ app/src/main/res/layout/activity_settings.xml
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
-        tools:context="org.secuso.example.SettingsActivity">
+        tools:context="org.secuso.privacyfriendlyexample.SettingsActivity">
 
         <include layout="@layout/toolbar"/>
 
@@ -28,7 +28,7 @@
             android:layout_marginTop="?attr/actionBarSize"
             android:weightSum="1">
 
-        <fragment android:name="org.secuso.example.SettingsActivity$GeneralPreferenceFragment"
+        <fragment android:name="org.secuso.privacyfriendlyexample.SettingsActivity$GeneralPreferenceFragment"
             android:id="@+id/settings_fragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"/>
