+++ app/src/main/res/layout/app_bar_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.secuso.example.MainActivity">
+    tools:context="org.secuso.privacyfriendlyexample.MainActivity">
 
     <include layout="@layout/toolbar"/>
