+++ app/src/main/res/layout/content_main.xml
@@ -11,7 +11,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="org.secuso.example.MainActivity"
+    tools:context="org.secuso.privacyfriendlyexample.MainActivity"
     tools:showIn="@layout/app_bar_main"
     android:orientation="vertical"
     style="@style/AlertDialog.AppCompat">
