+++ app/src/test/java/org/secuso/privacyfriendlyexample/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package org.secuso.example;
+package org.secuso.privacyfriendlyexample;
 
 import org.junit.Test;
